+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Add `gha-utils labels` command to dump bundled label configuration files.
+- Rely on `gha-utils release-prep` for release preparation steps.
 
 ## [5.1.0 (2026-01-23)](https://github.com/kdeldycke/workflows/compare/v5.0.1...v5.1.0)